Fix #17743 - [Merge] Merge 2023 nutriscore version
Conflicts:
	becpg-integration-runner/src/test/java/fr/becpg/test/repo/product/score/NutriScoreIT.java
	becpg-plm/becpg-plm-core/src/main/java/fr/becpg/repo/product/formulation/score/NutriScore.java
	becpg-plm/becpg-plm-core/src/main/java/fr/becpg/repo/product/formulation/score/NutriScoreContext.java
	becpg-plm/becpg-plm-core/src/main/java/fr/becpg/repo/product/helper/Nutrient5C2021Helper.java
	becpg-plm/becpg-plm-core/src/main/resources/alfresco/module/becpg-plm-core/becpg-config-plm.properties

+++ becpg-plm/becpg-plm-core/src/main/java/fr/becpg/repo/admin/patch/MigrateNutrientProfilePatch.java
@@ -8,6 +8,7 @@ import org.alfresco.model.ContentModel;
 import org.alfresco.repo.batch.BatchProcessWorkProvider;
 import org.alfresco.repo.batch.BatchProcessor;
 import org.alfresco.repo.batch.BatchProcessor.BatchProcessWorker;
+import org.alfresco.repo.batch.BatchProcessor.BatchProcessWorkerAdaptor;
 import org.alfresco.repo.domain.node.NodeDAO;
 import org.alfresco.repo.domain.patch.PatchDAO;
 import org.alfresco.repo.domain.qname.QNameDAO;
@@ -24,6 +25,7 @@ import org.springframework.extensions.surf.util.I18NUtil;
 
 import fr.becpg.model.BeCPGModel;
 import fr.becpg.model.NutrientProfileCategory;
+import fr.becpg.model.NutrientProfileVersion;
 import fr.becpg.model.PLMModel;
 import fr.becpg.repo.RepoConsts;
 
@@ -158,11 +160,6 @@ public class MigrateNutrientProfilePatch extends AbstractBeCPGPatch {
 				return result.size();
 			}
 			
-			@Override
-			public long getTotalEstimatedWorkSizeLong() {
-				return result.size();
-			}
-
 			public Collection<NodeRef> getNextWork() {
 
 				result.clear();
@@ -191,26 +188,21 @@ public class MigrateNutrientProfilePatch extends AbstractBeCPGPatch {
 				transactionService.getRetryingTransactionHelper(), workProvider, BATCH_THREAD, BATCH_SIZE,
 				applicationEventPublisher, logger, 1000);
 
-		BatchProcessWorker<NodeRef> worker = new BatchProcessWorker<NodeRef>() {
+		BatchProcessWorker<NodeRef> worker = new BatchProcessWorkerAdaptor<NodeRef>() {
 
-			public void afterProcess() throws Throwable {
-				ruleService.enableRules();
-			}
-
-			public void beforeProcess() throws Throwable {
-				ruleService.disableRules();
-			}
-
+			@Override
 			public String getIdentifier(NodeRef entry) {
 				return entry.toString();
 			}
 
+			@Override
 			public void process(NodeRef nutrientProfile) throws Throwable {
+				
 				AuthenticationUtil.setAdminUserAsFullyAuthenticatedUser();
 
-				policyBehaviourFilter.disableBehaviour();
-				
+				ruleService.disableRules();
 				integrityChecker.setEnabled(false);
+				policyBehaviourFilter.disableBehaviour();
 				
 				List<AssociationRef> sourceAssocs = nodeService.getSourceAssocs(nutrientProfile, ASSOC_NUTRIENT_PROFILE_REF);
 				
@@ -218,19 +210,29 @@ public class MigrateNutrientProfilePatch extends AbstractBeCPGPatch {
 					
 					String nutrientProfileClass = (String) nodeService.getProperty(nutrientProfile, BeCPGModel.PROP_CHARACT_NAME);
 					
-					boolean nutrientProfileClassKnown = nutrientProfileClass != null && (nutrientProfileClass.contains("Others") || nutrientProfileClass.contains("Beverages") || nutrientProfileClass.contains("Fats") || nutrientProfileClass.contains("Cheeses"));
+					boolean nutrientProfileClassKnown = nutrientProfileClass != null 
+							&& (nutrientProfileClass.contains("Others") ||
+									nutrientProfileClass.contains("Autres") 
+									|| nutrientProfileClass.contains("Beverages") 
+									|| nutrientProfileClass.contains("Boissons") 
+									|| nutrientProfileClass.contains("Fats") || nutrientProfileClass.contains("Matières grasses")
+									|| nutrientProfileClass.contains("Cheeses")
+									|| nutrientProfileClass.contains("Fromages"));
 					
 					if (nutrientProfileClassKnown) {
 						for (AssociationRef sourceAssoc : sourceAssocs) {
-							
-							if (nutrientProfileClass.contains("Others")) {
+							if (nutrientProfileClass.contains("Others") || nutrientProfileClass.contains("Autres")) {
 								nodeService.setProperty(sourceAssoc.getSourceRef(), PLMModel.PROP_NUTRIENT_PROFILE_CATEGORY, NutrientProfileCategory.Others.toString());
-							} else if (nutrientProfileClass.contains("Beverages")) {
+							} else if (nutrientProfileClass.contains("Beverages") || nutrientProfileClass.contains("Boissons")) {
 								nodeService.setProperty(sourceAssoc.getSourceRef(), PLMModel.PROP_NUTRIENT_PROFILE_CATEGORY, NutrientProfileCategory.Beverages.toString());
-							} else if (nutrientProfileClass.contains("Fats")) {
+							} else if (nutrientProfileClass.contains("Fats") || nutrientProfileClass.contains("Matières grasses")) {
 								nodeService.setProperty(sourceAssoc.getSourceRef(), PLMModel.PROP_NUTRIENT_PROFILE_CATEGORY, NutrientProfileCategory.Fats.toString());
-							} else if (nutrientProfileClass.contains("Cheeses")) {
+							} else if (nutrientProfileClass.contains("Cheeses") || nutrientProfileClass.contains("Fromages")) {
 								nodeService.setProperty(sourceAssoc.getSourceRef(), PLMModel.PROP_NUTRIENT_PROFILE_CATEGORY, NutrientProfileCategory.Cheeses.toString());
+							}
+							
+							if (nutrientProfileClass.contains(NutrientProfileVersion.VERSION_2023.toString())) {
+								nodeService.setProperty(sourceAssoc.getSourceRef(), PLMModel.PROP_NUTRIENT_PROFILE_VERSION, NutrientProfileVersion.VERSION_2023.toString());
 							}
 							
 							nodeService.removeAssociation(sourceAssoc.getSourceRef(), nutrientProfile, ASSOC_NUTRIENT_PROFILE_REF);
@@ -240,7 +242,7 @@ public class MigrateNutrientProfilePatch extends AbstractBeCPGPatch {
 				}
 				
 				policyBehaviourFilter.enableBehaviour();
-				
+				ruleService.enableRules();
 				integrityChecker.setEnabled(true);
 			}
 
@@ -249,20 +251,26 @@ public class MigrateNutrientProfilePatch extends AbstractBeCPGPatch {
 		batchProcessor.processLong(worker, true);
 		
 		NodeRef companyHomeNodeRef = repository.getCompanyHome();
-		
-		NodeRef systemNodeRef = repoService.getFolderByPath(companyHomeNodeRef, RepoConsts.PATH_SYSTEM);
-		
-		NodeRef charactsNodeRef = repoService.getFolderByPath(systemNodeRef, RepoConsts.PATH_CHARACTS);
-		
-		NodeRef entityListNodeRef = repoService.getFolderByPath(charactsNodeRef, "bcpg:entityLists");
-		
-		NodeRef nutrientProfilesCategoryNodeRef = repoService.getFolderByPath(entityListNodeRef, PATH_NUTRIENTPROFILES);
-
-		if (nutrientProfilesCategoryNodeRef != null && nodeService.exists(nutrientProfilesCategoryNodeRef)) {
-			nodeService.addAspect(nutrientProfilesCategoryNodeRef, ContentModel.ASPECT_TEMPORARY, null);
-			nodeService.deleteNode(nutrientProfilesCategoryNodeRef);
+		if (companyHomeNodeRef != null) {
+			NodeRef systemNodeRef = repoService.getFolderByPath(companyHomeNodeRef, RepoConsts.PATH_SYSTEM);
+			if (systemNodeRef != null) {
+				NodeRef charactsNodeRef = repoService.getFolderByPath(systemNodeRef, RepoConsts.PATH_CHARACTS);
+				if (charactsNodeRef != null) {
+					NodeRef entityListNodeRef = repoService.getFolderByPath(charactsNodeRef, "bcpg:entityLists");
+					if (entityListNodeRef != null) {
+						NodeRef nutrientProfilesCategoryNodeRef = repoService.getFolderByPath(entityListNodeRef, PATH_NUTRIENTPROFILES);
+						if (nutrientProfilesCategoryNodeRef != null && nodeService.exists(nutrientProfilesCategoryNodeRef)) {
+							if (nodeService.getChildAssocs(nutrientProfilesCategoryNodeRef).isEmpty()) {
+								nodeService.addAspect(nutrientProfilesCategoryNodeRef, ContentModel.ASPECT_TEMPORARY, null);
+								nodeService.deleteNode(nutrientProfilesCategoryNodeRef);
+								logger.info("deleting nutrient profile list: " + nutrientProfilesCategoryNodeRef);
+							}
+						}
+					}
+				}
+			}
 		}
-
+		
 		return I18NUtil.getMessage(MSG_SUCCESS);
 	}
